# Conflicts:
#	.gitignore
#	mmm_audio/Analysis.mojo
#	mmm_audio/Buffer_Module.mojo
#	mmm_python/MMMAudio.py
#	pixi.lock
#	pixi.toml
#	testing/MFCC_Validation.mojo
#	testing/MFCC_Validation.py
#	testing/MelBands_Validation.mojo
#	testing/MelBands_Validation.py
#	testing/RMS_Validation.py
#	testing/SpectralCentroid_Validation.mojo
#	testing/SpectralCentroid_Validation.py
#	testing/SpectralCentroid_Validation.scd
#	testing/SpectralCrest_Validation.mojo
#	testing/SpectralCrest_Validation.py
#	testing/SpectralFlatness_Validation.mojo
#	testing/SpectralFlatness_Validation.py
#	testing/SpectralKurtosis_Validation.mojo
#	testing/SpectralKurtosis_Validation.py
#	testing/SpectralRolloff_Validation.mojo
#	testing/SpectralRolloff_Validation.py
#	testing/SpectralShape_Validation.scd
#	testing/SpectralSkewness_Validation.mojo
#	testing/SpectralSkewness_Validation.py
#	testing/SpectralSpread_Validation.mojo
#	testing/SpectralSpread_Validation.py
#	testing/YIN_Validation.mojo
#	testing/YIN_Validation.py
#	validation/run_all.py

+++ mmm_audio/Oversampling_Module.mojo
@@ -1,5 +1,4 @@
-from .Filters import lpf_LR4
-from .MMMWorld_Module import *
+from mmm_audio import *
 
 struct Oversampling[num_chans: Int = 1, times_oversampling: Int = 0](Representable, Movable, Copyable):
     """A struct that collects ` times_oversampling` samples and then downsamples them using a low-pass filter. Add a sample for each oversampling iteration with `add_sample()`, then get the downsampled output with `get_sample()`.
@@ -10,7 +9,7 @@ struct Oversampling[num_chans: Int = 1, times_oversampling: Int = 0](Representab
     """
 
     var buffer: InlineArray[SIMD[DType.float64, Self.num_chans], Self.times_oversampling]  # Buffer for oversampled values
-    var counter: Int64
+    var counter: Int
     var lpf: OS_LPF4[Self.num_chans]
 
     fn __init__(out self, world: World):
